backend/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
PyMySQL
python-dotenv
Werkzeug
PyJWT
cryptography
pandas
openpyxl
